break out General settings tab code
